refactor: update source
Move `socialLinks` to `links` constant file
Update structure of `monthMap`
Slightly reformat middleware and next config files
