+++ src/constants/links.ts
@@ -1,0 +1,13 @@
+import {
+  FacebookIcon,
+  InstagramIcon,
+  LinkedInIcon,
+  TwitterIcon,
+} from "@/components/icons";
+
+export const socialLinks = [
+  { id: "facebook", icon: FacebookIcon, href: "https://facebook.com/" },
+  { id: "twitter", icon: TwitterIcon, href: "https://twitter.com/" },
+  { id: "instagram", icon: InstagramIcon, href: "https://instagram.com/" },
+  { id: "linkedin", icon: LinkedInIcon, href: "https://linkedin.com/" },
+];
