+++ src/lib/links.ts
@@ -1,9 +1,3 @@
-import {
-  FacebookIcon,
-  InstagramIcon,
-  LinkedInIcon,
-  TwitterIcon,
-} from "@/components/icons";
 import { routes } from "@/constants/routes";
 import { type DictionaryType } from "@/locale/get-dictionary";
 
@@ -49,10 +43,3 @@ export const getLinks = (
 
   return baseLinks;
 };
-
-export const socialLinks = [
-  { id: "facebook", icon: FacebookIcon, href: "https://facebook.com/" },
-  { id: "twitter", icon: TwitterIcon, href: "https://twitter.com/" },
-  { id: "instagram", icon: InstagramIcon, href: "https://instagram.com/" },
-  { id: "linkedin", icon: LinkedInIcon, href: "https://linkedin.com/" },
-];
